requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
eventlet==0.33.3
numpy<2.0
Pillow==10.0.1
opencv-python-headless==4.8.1.78
python-dateutil==2.8.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.0
scikit-image==0.21.0
joblib==1.5.1
google-generativeai==0.3.2